Add set-region command (#15)
In case a user uses the `current-shift` command, it might be possible that
the tool has no reference what the current shift is based on the region. To
allow the user to explicitly set a region, a convenience command should be
added to the tool.

Add function to update config file with new data.

Add detection to guess the current shift based on timezone.

Update `current-shift` command to include helper message on how to
configure the current user region.

Introduce constant for command name.

Introduce ShiftTime type to simplify code.

Add new func `getDataFromYAMLFile`.

Change func `ChangeYAMLFile` to be able to modify every entry in the `.pd.yml` file.